Performance: Calc colors only when changed. 🚀
This fixes the perf bug with a laggy UI.

However, I suppose that the egui framework has a better method
to achieve the same result.

In my approach, I store a hash of the settings and draw
into the pixel buffer only when the settings have changed.

The effect is a noticeable improvement in the UI.  🙂